Move Ax+Y into implementation file
