+++ src/tensors/tensors.cpp
@@ -1,0 +1,88 @@
+//========================================================================================
+// (C) (or copyright) 2025. Triad National Security, LLC. All rights reserved.
+//
+// This program was produced under U.S. Government contract 89233218CNA000001 for Los
+// Alamos National Laboratory (LANL), which is operated by Triad National Security, LLC
+// for the U.S. Department of Energy/National Nuclear Security Administration. All rights
+// in the program are reserved by Triad National Security, LLC, and the U.S. Department
+// of Energy/National Nuclear Security Administration. The Government is granted for
+// itself and others acting on its behalf a nonexclusive, paid-up, irrevocable worldwide
+// license in this material to reproduce, prepare derivative works, distribute copies to
+// the public, perform publicly and display publicly, and to permit others to do so.
+//========================================================================================
+
+#include <vector>
+
+#include "tensors/tensors.hpp"
+
+namespace parthenon {
+namespace tensors {
+
+TensorTrain aXPlusY(pool_map_t &pool_map, const Real a, const TensorTrain &X,
+                    const TensorTrain &Y) {
+  PARTHENON_REQUIRE_THROWS(X.GetNumCores() == Y.GetNumCores(),
+                           "Ensure tensor"
+                           " trains being added have same number of cores");
+
+  // declare and construct the tensor cores for the resulting train
+  std::vector<TensorCoreHost> cores;
+  for (int i = 0; i < X.GetNumCores(); i++) {
+    PARTHENON_REQUIRE_THROWS(X.GetPhysicalIndexSize(i) == Y.GetPhysicalIndexSize(i),
+                             "Ensure tensor trains being added have"
+                             " same physical index size in corresponding cores");
+
+    const std::size_t rL = (i == 0) ? 1 : X.GetLeftRank(i) + Y.GetLeftRank(i);
+    const std::size_t nc = X.GetPhysicalIndexSize(i);
+    const std::size_t rR =
+        (i == X.GetNumCores() - 1) ? 1 : X.GetRightRank(i) + Y.GetRightRank(i);
+
+    cores.emplace_back(pool_map, rL, nc, rR);
+  }
+
+  // construct the train that will contain the result with these cores
+  TensorTrain Z(std::to_string(a) + "*" + X.label() + " + " + Y.label(), cores);
+
+  auto Xcores = X.cores_device_;
+  auto Ycores = Y.cores_device_;
+  auto Zcores = Z.cores_device_;
+  par_for(
+      PARTHENON_AUTO_LABEL, 0, X.GetNumCores() - 1, KOKKOS_LAMBDA(const int i) {
+        const std::size_t nXL = Xcores(i).GetLeftRank();
+        const std::size_t nXR = Xcores(i).GetRightRank();
+
+        // zero initialize
+        for (int iL = 0; iL < Zcores(i).GetLeftRank(); iL++) {
+          for (int iR = 0; iR < Zcores(i).GetRightRank(); iR++) {
+            for (int ic = 0; ic < Zcores(i).GetPhysicalIndexSize(); ic++) {
+              Zcores(i)(iL, ic, iR) = 0.;
+            }
+          }
+        }
+
+        // from X
+        for (int iL = 0; iL < nXL; iL++) {
+          for (int iR = 0; iR < nXR; iR++) {
+            for (int ic = 0; ic < Xcores(i).GetPhysicalIndexSize(); ic++) {
+              Zcores(i)(iL, ic, iR) = Xcores(i)(iL, ic, iR);
+              if (i == 0) Zcores(i)(iL, ic, iR) *= a;
+            }
+          }
+        }
+
+        // from Y - left (right) offset is zero for first (last) core
+        const int oL = (i > 0) * nXL;
+        const int oR = (i < X.GetNumCores() - 1) * nXR;
+        for (int iL = 0; iL < Ycores(i).GetLeftRank(); iL++) {
+          for (int iR = 0; iR < Ycores(i).GetRightRank(); iR++) {
+            for (int ic = 0; ic < Ycores(i).GetPhysicalIndexSize(); ic++) {
+              Zcores(i)(oL + iL, ic, oR + iR) = Ycores(i)(iL, ic, iR);
+            }
+          }
+        }
+      });
+
+  return Z;
+} // AXPlusY
+
+} // namespace tensors
+} // namespace parthenon
